De-bias the rng by always using 2x64 bit state in xorshift128+
Really, this was the intention, as the '128' means 128 bits of state. To
use on smaller types (e.g. uint32_t) it's likely the A, B and C values
need changing, as currently it makes a pretty crappy rng due to the
shifts clamping

+++ game/state/gamestate_serialize.cpp
@@ -135,7 +135,7 @@ void serializeIn(const GameState *state, sp<SerializationNode> node, Xorshift128
 	if (!node)
 		return;
 
-	uint32_t s[2] = {0, 0};
+	uint64_t s[2] = {0, 0};
 	serializeIn(state, node->getNode("s0"), s[0]);
 	serializeIn(state, node->getNode("s1"), s[1]);
 	t.setState(s);
@@ -234,8 +234,8 @@ void serializeOut(sp<SerializationNode> node, const Xorshift128Plus<uint32_t> &t
 	if (!node)
 		return;
 
-	uint32_t s[2] = {0, 0};
-	uint32_t sr[2] = {0, 0};
+	uint64_t s[2] = {0, 0};
+	uint64_t sr[2] = {0, 0};
 	t.getState(s);
 	ref.getState(sr);
 	serializeOut(node->addNode("s0"), s[0], sr[0]);
